requirements.txt
discord.py[voice]==2.3.2
dnspython==2.4.2
mongoengine==0.27.0
pymongo==4.6.0
python-dotenv==1.0.0
